Nil check `tech.unit` before attempting use as it is `nil` on trigger technologies.
As per https://lua-api.factorio.com/latest/prototypes/TechnologyPrototype.html#unit `tech.unit` is optional (specifically it is nil when it is a trigger technology) and must be `nil` checked prior to use.

Also added a log message in the event this case is hit.  Unable to find any other `log` in this mod so uncertain what is the preferred format, so just followed a general format.